VR Controller working(!)
The 'Press' function in the 'InputTester' script (which is in the right controller object) works as intended. Now we just have to link the press to the interaction function, I think it might work. Otherwise the timer buffer has to be put inside the interaction scripts too, which would be good either way.